Redesign strucutre around suitability, membership and discrete functions
